refactor: break up addr module into file
